posix: extract into module/posix/ following the _socket pattern
Carries the shared stat_result_type helper.  host module is bumped to
pub(crate) so interp_posix can route through it.  importing.rs shrinks
by 2594 lines.
