+++ pyre/pyre-interpreter/src/module/posix/moduledef.rs
@@ -1,0 +1,7 @@
+//! posix module definition — PyPy: pypy/module/posix/moduledef.py
+
+use crate::DictStorage;
+
+pub fn init(ns: &mut DictStorage) {
+    super::interp_posix::register_module(ns);
+}
